Guesser game assignment
created a method called register where players wills register to play the game. once registered they will login with the registered credentials  for that login method was created and if login details does not match with the registered details login will not happen and also the guessing numbers should be less than 10.